Add check for area and perimeter of triangle and circle

+++ src/libgeometry/geo.h
@@ -1,12 +1,12 @@
 #ifndef GEO_H
 #define GEO_H
 typedef struct Point {
-    int x;
-    int y;
+    double x;
+    double y;
 } point;
 typedef struct Vector {
-    int a;
-    int b;
+    double a;
+    double b;
 } vector;
 
 vector funVector(point first, point second);
